# swagger/openapi.yaml
...
openapi: 3.0.0
servers:
- url: http://localhost:3000/
info:
  description: 'This is a sample E-Commerce app.  This project requires that you build a fully-functioning e-commerce application REST API that allows users to perform various CRUD operations such as registering an account, browsing products for sale, etc.'
  version: 1.0.0
  title: E-Commerce REST API application
  termsOfService: https://www.codecademy.com/about
  contact:
    email: easoftware@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
  "/users":
    post:
      tags:
      - users
      summary: Create user
      description: This can only be done when the user registers.
      operationId: createUser
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '400':
          description: email already used, required value missing
        '404':
          description: user not inserted
      requestBody:
        "$ref": "#/components/requestBodies/UsersPost"
    get:
      tags:
      - users
      summary: Gets all users
      description: ''
      operationId: getAllUsers
      responses:
        '200':
          description: successful opertion
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Users"  
        '404':
          description: 'no user rows found'
  "/users/{id}":
    get:
      tags:
      - users
      summary: Get user by user id
      description: ''
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: 'The user id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '404':
          description: Invalid User ID supplied, User not found
    put:
      tags:
      - users
      summary: update user by user id
      description: 'This can only be done by the logged in user.'
      operationId: updateUserById
      parameters:
      - name: id
        in: path
        description: 'The product id that needs to be updated. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
        '400':
          description: email already used, required value missing
        '404':
          description: Invalid User ID supplied, User not found
      requestBody:
        "$ref": "#/components/requestBodies/UsersPut"
    delete:
      tags:
      - users
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: 'The user id that needs to be deleted. In testing, use before() to generate user to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted user - user's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: User not found
        '409':
          description: Cannot delete - constraint error
...
  "/products":
    post:
      tags:
      - products
      description: Can only be done by admin
      operationId: createProduct
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Products"
        '400': 
          description: name already used, model_number already used, required value missing
        '404':
          description: Product not inserted
      requestBody:
        "$ref": "#/components/requestBodies/Products"
    get:
      tags:
      - products
      description: Can only be done by admin
      operationId: getProduct
      responses:
        '200':
          description: successful opertion
          content:
            application/json:
              schema: 
                type: array
                items: 
                  "$ref": "#/components/schemas/Products"  
        '404':
          description: 'no product rows found'
  "/products/{id}":  
    get:
      tags:
      - products
      summary: get procuct by product id
      description: ''
      operationId: getProductById
      parameters:
      - name: id
        in: path
        description: 'The product id that needs to be fetched. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Products"
        '404':
          description: Invalid Product ID supplied, Product not found
    put:
      tags:
      - products
      summary: update product by product id
      description: Can only be done by admin
      operationId: updateProductById
      parameters:
      - name: id
        in: path
        description: 'The product id that needs to be updated. Use 2 for testing. '
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Products"
        '400':
          description: name already used, model_number already used, required value missing
        '404':
          description: Invalid Product ID supplied, Product not found
      requestBody:
        "$ref": "#/components/requestBodies/Products"
    delete:
      tags:
      - products
      summary: Delete product
      description: Can only be done by admin
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        description: 'The product id that needs to be deleted. In testing, use before() to generate product to delete.'
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: Deleted product - product's id returned as a string
          content:
            text/plain: 
              schema:
                type: string
        '404':
          description: Product not found
        '409':
          description: Cannot delete - constraint error
...
components:
  schemas:
...
    Products:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        model_number:
          type: string
        description:
          type: string
        price:
          type: number
          format: Decimal
...
    Users:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
          format: email
        password_hash:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
      xml:
        name: User
...
  requestBodies:
...
    Products:
      content:
        application/json:
          schema:
            type: object
            properties:
              name:
                type: string
              model_number:
                type: string
              description:
                type: string
              price:
                type: number
                format: Decimal
...
    UsersPost:
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone:
                type: string
      description: 'Note: password is sent in plaintext, and password_hash is in Response'
      required: true
    UsersPut:
      content:
        application/json:
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              first_name:
                type: string
              last_name:
                type: string
              phone:
                type: string
      description: ''
      required: true
...